wrapping up section 8

+++ resources/item.py
@@ -1,7 +1,7 @@
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
 from sqlalchemy.exc import SQLAlchemyError
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt
 
 from db import db
 from models.item import ItemModel
@@ -16,8 +16,13 @@ class Item(MethodView):
     def get(self, item_id):
         return ItemModel.query.get_or_404(item_id)
 
-    @jwt_required()
+    @jwt_required(fresh=True)
     def delete(self, item_id):
+        # Esse trecho é um exemplo de uso do método add_claims_to_jwt
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required.")
+
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
@@ -50,7 +55,7 @@ class ItemList(MethodView):
         # Não precisamos mais fazer o cast para uma list pq Schema fará
         return ItemModel.query.all()
 
-    @jwt_required()
+    @jwt_required(fresh=True)
     @blp.arguments(ItemSchema)
     @blp.response(201, ItemSchema)
     def post(self, item_data):
